refactor: Controller의 역할을 Application으로 이동 & 비즈니스로직을 담당하는  RacingGame 객체 분리
